Fix phan-taint-check false positives
One issue due to array taint conflation ($options['id'] was used,
which was fine, but other members of the $options array weren't and
the plugin cannot track individual array members)

The rest are false positives from the plugin not being able to
determine that a page_id fetched from the db will be an integer
and safe to directly use in an SQL query.

Bug: T202375

+++ maintenance/purgeReviewablePages.php
@@ -66,8 +66,8 @@ class PurgeReviewablePages extends Maintenance {
 		}
 		# Do remaining chunk
 		$end += $this->mBatchSize - 1;
-		$blockStart = $start;
-		$blockEnd = $start + $this->mBatchSize - 1;
+		$blockStart = (int)$start;
+		$blockEnd = (int)( $start + $this->mBatchSize - 1 );
 
 		$count = 0;
 		while ( $blockEnd <= $end ) {
