+++ maintenance/reviewAllPages.php
@@ -48,8 +48,8 @@ class ReviewAllPages extends Maintenance {
 		}
 		# Do remaining chunk
 		$end += $this->mBatchSize - 1;
-		$blockStart = $start;
-		$blockEnd = $start + $this->mBatchSize - 1;
+		$blockStart = (int)$start;
+		$blockEnd = (int)( $start + $this->mBatchSize - 1 );
 		$count = 0;
 		$changed = 0;
 		$flags = FlaggedRevs::quickTags( FR_CHECKED ); // Assume basic level
